Form data adding to database and added validation
I added the form data and made it so that the users input is added to
the database, just like the user credentials are when signing up. I also
added some validation so that the username provided by the user before
filling out the quiz must exist and match the logged in user.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>